+++ apps/Standalone/src/templates/app/LocalTemplates.tsx
@@ -26,9 +26,9 @@ const loadLocalTemplateFromResourcePath = async (resourcePath: string, artifactT
     : (await import(`./../../../../../__mocks__/templates//${resourcePath}/${artifactType}.json`)).default;
 };
 
-const localTemplateManifestPaths = ['BasicWorkflowOnly', 'SimpleConnectionParameter', 'SimpleAccelerator'];
+const localTemplateManifestPaths = ['BasicWorkflowOnly', 'SimpleConnectionParameter', 'SimpleAccelerator', 'SimpleParametersOnly'];
 
-export const LocalTemplatesStandalone = () => {
+export const LocalTemplates = () => {
   const theme = useSelector((state: RootState) => state.workflowLoader.theme);
   const { hostingPlan } = useSelector((state: RootState) => state.workflowLoader);
   const { data: localManifests } = useQuery(
@@ -67,6 +67,7 @@ export const LocalTemplatesStandalone = () => {
         connectionReferences={{}}
         services={services}
         isConsumption={isConsumption}
+        isCreateView={!isConsumption}
         customTemplates={localManifests}
         existingWorkflowName={undefined}
       >
@@ -204,8 +205,9 @@ const getServices = (isConsumption: boolean, getLocalResource: (resourcePath: st
         openBladeAfterCreate: (_workflowName: string | undefined) => {
           window.alert('Open blade after create, consumption creation is complete');
         },
-        onAddBlankWorkflow: () => {
-          console.log('On add blank workflow click');
+        onAddBlankWorkflow: async () => {
+          await new Promise((resolve) => setTimeout(resolve, 1000));
+          window.alert('On Blank Workflow Click');
         },
         getCustomResource: getLocalResource,
       })
@@ -220,8 +222,8 @@ const getServices = (isConsumption: boolean, getLocalResource: (resourcePath: st
         openBladeAfterCreate: (workflowName: string | undefined) => {
           window.alert(`Open blade after create, workflowName is: ${workflowName}`);
         },
-        onAddBlankWorkflow: () => {
-          console.log('On add blank workflow click');
+        onAddBlankWorkflow: async () => {
+          window.alert('On Blank Workflow Click');
         },
         getCustomResource: getLocalResource,
       });