+++ libs/designer/src/lib/core/utils/parameters/helper.ts
@@ -117,7 +117,6 @@ import {
   isBodySegment,
   canStringBeConverted,
   splitAtIndex,
-  unescapeString,
 } from '@microsoft/logic-apps-shared';
 import type {
   AuthProps,
@@ -163,6 +162,7 @@ import type {
   OpenAPIV2,
   OperationManifest,
   RecurrenceSetting,
+  OperationInfo,
 } from '@microsoft/logic-apps-shared';
 import { createAsyncThunk, type Dispatch } from '@reduxjs/toolkit';
 import { getInputDependencies } from '../../actions/bjsworkflow/initialize';
@@ -1773,7 +1773,11 @@ export const updateParameterAndDependencies = createAsyncThunk(
     } = actionPayload;
     const parameter = nodeInputs.parameterGroups[groupId].parameters.find((param) => param.id === parameterId) ?? {};
     const updatedParameter = { ...parameter, ...properties } as ParameterInfo;
-    updatedParameter.validationErrors = validateParameter(updatedParameter, updatedParameter.value);
+    updatedParameter.validationErrors = validateParameter(
+      updatedParameter,
+      updatedParameter.value,
+      shouldEncodeParameterValueForOperationBasedOnMetadata(operationInfo)
+    );
     const propertiesWithValidations = { ...properties, validationErrors: updatedParameter.validationErrors };
 
     const parametersToUpdate = [
@@ -2517,7 +2521,8 @@ function fetchErrorWhenDependenciesNotReady(
 function getStringifiedValueFromEditorViewModel(
   parameter: ParameterInfo,
   isDefinitionValue: boolean,
-  idReplacements?: Record<string, string>
+  idReplacements?: Record<string, string>,
+  shouldEncodeBasedOnMetadata = true
 ): string | undefined {
   const { editor, editorOptions, editorViewModel } = parameter;
   switch (editor?.toLowerCase()) {
@@ -2532,12 +2537,14 @@ function getStringifiedValueFromEditorViewModel(
           const keyValue = parameterValueToString(
             { type: types[0], value: item.key, ...commonProperties } as any,
             isDefinitionValue,
-            idReplacements
+            idReplacements,
+            shouldEncodeBasedOnMetadata
           );
           const valueValue = parameterValueToString(
             { type: types[1], value: item.value, ...commonProperties } as any,
             isDefinitionValue,
-            idReplacements
+            idReplacements,
+            shouldEncodeBasedOnMetadata
           );
 
           if (keyValue || valueValue) {
@@ -2552,7 +2559,16 @@ function getStringifiedValueFromEditorViewModel(
     case constants.EDITOR.CONDITION:
       return editorOptions?.isOldFormat
         ? iterateSimpleQueryBuilderEditor(editorViewModel.itemValue, editorViewModel.isRowFormat, idReplacements)
-        : JSON.stringify(recurseSerializeCondition(parameter, editorViewModel.items, isDefinitionValue, idReplacements));
+        : JSON.stringify(
+            recurseSerializeCondition(
+              parameter,
+              editorViewModel.items,
+              isDefinitionValue,
+              idReplacements,
+              /* errors */ undefined,
+              shouldEncodeBasedOnMetadata
+            )
+          );
     case constants.EDITOR.FLOATINGACTIONMENU: {
       if (!editorViewModel || editorOptions?.menuKind !== FloatingActionMenuKind.outputs) {
         return undefined;
@@ -2633,7 +2649,8 @@ export const recurseSerializeCondition = (
   editorViewModel: any,
   isDefinitionValue: boolean,
   idReplacements?: Record<string, string>,
-  errors?: string[]
+  errors?: string[],
+  shouldEncodeBasedOnMetadata = true
 ): any => {
   const returnVal: any = {};
   const commonProperties = { supressCasting: parameter.suppressCasting, info: parameter.info };
@@ -2652,12 +2669,14 @@ export const recurseSerializeCondition = (
     const operand1String = parameterValueToString(
       { type: 'any', value: operand1, ...commonProperties } as any,
       isDefinitionValue,
-      idReplacements
+      idReplacements,
+      shouldEncodeBasedOnMetadata
     );
     const operand2String = parameterValueToString(
       { type: 'any', value: operand2, ...commonProperties } as any,
       isDefinitionValue,
-      idReplacements
+      idReplacements,
+      shouldEncodeBasedOnMetadata
     );
     if (errors && errors.length === 0 && (operand1String || operand2String)) {
       if (!operand1String) {
@@ -2695,7 +2714,7 @@ export const recurseSerializeCondition = (
       ];
     }
     returnVal[condition] = items.map((item: any) => {
-      return recurseSerializeCondition(parameter, item, isDefinitionValue, idReplacements, errors);
+      return recurseSerializeCondition(parameter, item, isDefinitionValue, idReplacements, errors, shouldEncodeBasedOnMetadata);
     });
   }
   return returnVal;
@@ -3497,7 +3516,8 @@ export function getInterpolatedExpression(expression: string, parameterType: str
 export function parameterValueToString(
   parameterInfo: ParameterInfo,
   isDefinitionValue: boolean,
-  idReplacements?: Record<string, string>
+  idReplacements?: Record<string, string>,
+  shouldEncodeBasedOnMetadata = true
 ): string | undefined {
   const { value: remappedValue, didRemap } = isRecordNotEmpty(idReplacements)
     ? remapValueSegmentsWithNewIds(parameterInfo.value, idReplacements ?? {})
@@ -3524,13 +3544,20 @@ export function parameterValueToString(
     }
   }
 
-  const valueFromEditor = getStringifiedValueFromEditorViewModel(remappedParameterInfo, isDefinitionValue, idReplacements);
+  const valueFromEditor = getStringifiedValueFromEditorViewModel(
+    remappedParameterInfo,
+    isDefinitionValue,
+    idReplacements,
+    shouldEncodeBasedOnMetadata
+  );
   if (valueFromEditor !== undefined) {
     return valueFromEditor;
   }
 
   const parameter = { ...remappedParameterInfo };
-  const requiresUrlEncoding = parameter.info.in === ParameterLocations.Path || parameter.info.encode !== undefined;
+  const requiresUrlEncoding = shouldEncodeBasedOnMetadata
+    ? parameter.info.in === ParameterLocations.Path || parameter.info.encode !== undefined
+    : parameter.info.in === ParameterLocations.Path;
   const value = parameter.value.filter((segment) => segment.value !== '');
 
   if (!value || !value.length) {
@@ -3763,14 +3790,13 @@ function parameterValueToStringWithoutCasting(value: ValueSegment[], forValidati
   const shouldInterpolateTokens = (value.length > 1 || shouldInterpolateSingleToken) && value.some(isTokenValueSegment);
 
   return value
-    .map((segment) => {
-      const { value: segmentValue } = segment;
-      if (isTokenValueSegment(segment)) {
-        const token = forValidation ? segmentValue || null : unescapeString(segmentValue);
-        return shouldInterpolateTokens ? `@{${token}}` : `@${token}`;
+    .map((expression) => {
+      let expressionValue = forValidation ? expression.value || null : expression.value;
+      if (isTokenValueSegment(expression)) {
+        expressionValue = shouldInterpolateTokens ? `@{${expressionValue}}` : `@${expressionValue}`;
       }
 
-      return forValidation ? segmentValue || null : segmentValue;
+      return expressionValue;
     })
     .join('');
 }
@@ -3949,14 +3975,15 @@ export function isParameterRequired(parameterInfo: ParameterInfo): boolean {
 export function validateParameter(
   parameter: ParameterInfo,
   parameterValue: ValueSegment[],
-  shouldValidateUnknownParameterAsError = false
+  shouldValidateUnknownParameterAsError = false,
+  shouldEncodeBasedOnMetadata = true
 ): string[] {
   const parameterType = getInferredParameterType(parameterValue, parameter.type);
   const parameterValueString = parameterValueToStringWithoutCasting(parameterValue, /* forValidation */ true);
   const isJsonObject = parameterType === constants.SWAGGER.TYPE.OBJECT;
 
   return isJsonObject
-    ? validateJSONParameter(parameter, parameterValue)
+    ? validateJSONParameter(parameter, parameterValue, shouldEncodeBasedOnMetadata)
     : validateStaticParameterInfo(parameter, parameterValueString, shouldValidateUnknownParameterAsError);
 }
 
@@ -4019,4 +4046,9 @@ export function validateUntilAction(
       })
     );
   }
+}
+
+export function shouldEncodeParameterValueForOperationBasedOnMetadata(operationInfo: OperationInfo): boolean {
+  const { connectorId } = operationInfo ?? {};
+  return !connectorId?.toLowerCase().includes('commondataservice');
 }