+++ libs/designer/src/lib/ui/panel/nodeDetailsPanel/tabs/parametersTab/index.tsx
@@ -26,6 +26,7 @@ import {
   parameterValueToString,
   remapEditorViewModelWithNewIds,
   remapValueSegmentsWithNewIds,
+  shouldEncodeParameterValueForOperationBasedOnMetadata,
   shouldUseParameterInGroup,
   updateParameterAndDependencies,
 } from '../../../../../core/utils/parameters/helper';
@@ -472,7 +473,8 @@ const ParameterSection = ({
                 suppressCasting,
               } as ParameterInfo,
               false,
-              idReplacements
+              idReplacements,
+              shouldEncodeParameterValueForOperationBasedOnMetadata(operationInfo)
             ) ?? '',
           getTokenPicker: (
             editorId: string,